separated Cart into smaller components
